add confirm dialog, fix mouse clicked on table, fix insert on table, etc

+++ src/main/server/frmMobil.java
@@ -396,29 +396,35 @@ public class frmMobil extends javax.swing.JFrame {
 
     private void BTN_HAPUSActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_BTN_HAPUSActionPerformed
         // TODO add your handling code here:
-//        id=String.valueOf(Txt_id_mobil.getText());
-        try {
-            // TODO add your handling code here:
-            id=String.valueOf(txt_id_mobil.getText());
-            sql="DELETE FROM tb_mobil"
-                    +" where id_mobil='"+ id +"'";
-            st=con.createStatement();
-            st.execute(sql);
-            bersih();
-            TampilData(sql);
-            JOptionPane.showMessageDialog(null, "Data Berhasil Dihapus");
-        } catch (Exception ex) {
-            JOptionPane.showMessageDialog(null, "Data Gagal Dihapus"+ex.getMessage());
-        }
-       
+        
+        id=String.valueOf(txt_id_mobil.getText());
+        int response = JOptionPane.showConfirmDialog(null, "Hapus Data ?", "Konfirmasi",
+        JOptionPane.YES_NO_OPTION, JOptionPane.QUESTION_MESSAGE);
+        if (response == JOptionPane.YES_OPTION) {
+                try {
+                    sql="DELETE FROM tb_mobil"
+                            +" where id_mobil='"+ id +"'";
+                    st=con.createStatement();
+                    st.execute(sql);
+                    bersih();
+                    TampilData(sql);
+                    JOptionPane.showMessageDialog(null, "Data Berhasil Dihapus");
+                } catch (Exception ex) {
+                    JOptionPane.showMessageDialog(null, "Data Gagal Dihapus"+ex.getMessage());
+                }
+            } 
     }//GEN-LAST:event_BTN_HAPUSActionPerformed
 
     private void BTN_KELUARActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_BTN_KELUARActionPerformed
         // TODO add your handling code here:
-        login masuk = new login();
-        masuk.setLocationRelativeTo(null);
-        masuk.setVisible(true);
-        dispose();
+        int response = JOptionPane.showConfirmDialog(null, "Keluar ?", "Konfirmasi",
+        JOptionPane.YES_NO_OPTION, JOptionPane.QUESTION_MESSAGE);
+           if (response == JOptionPane.YES_OPTION) {
+                 login masuk = new login();
+                 masuk.setLocationRelativeTo(null);
+                 masuk.setVisible(true);
+                 dispose();
+            } 
     }//GEN-LAST:event_BTN_KELUARActionPerformed
 
     private void Txt_nama_mobilActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_Txt_nama_mobilActionPerformed
